2.0

+++ src/main/java/co/uis/proyecto/entidades/Paciente.java
@@ -3,11 +3,16 @@ package co.uis.proyecto.entidades;
 import jakarta.persistence.*;
 import lombok.*;
 
-@Entity@Table(name="paciente")
-@Data@NoArgsConstructor@AllArgsConstructor@Builder
+@Entity
+@Table(name = "paciente")
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
 public class Paciente {
-    @Id@Column(name = "cedula")
-    private int cedula;
+    @Id
+    @Column(name = "cedula")
+    private Long cedula;
 
     @Column(name = "nombre")
     private String nombre;
