+++ src/main/java/co/uis/proyecto/entidades/Registro.java
@@ -2,14 +2,20 @@ package co.uis.proyecto.entidades;
 
 import jakarta.persistence.*;
 import lombok.*;
-@Entity@Table(name="registro")
-@Data@NoArgsConstructor@AllArgsConstructor@Builder
+
+@Entity
+@Table(name = "registro")
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
 public class Registro {
-    @Id@Column(name = "num_registro")
+    @Id
+    @Column(name = "num_registro")
     private int num_registro;
 
     @Column(name = "id_usuario")
-    private int id_usuario;
+    private Long id_usuario;
 
     @Column(name = "num_episodio")
     private int num_episodio;
