+++ src/main/java/co/uis/proyecto/repositorios/PacienteRepository.java
@@ -3,5 +3,5 @@ package co.uis.proyecto.repositorios;
 import co.uis.proyecto.entidades.Paciente;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PacienteRepository extends JpaRepository<Paciente,Integer> {
+public interface PacienteRepository extends JpaRepository<Paciente, Long> {
 }
